agents/steve/requirements.txt
# API Modules
fastapi[standard]
uvicorn>=0.34.2

# LLM Modules
langchain>=0.3.23
langchain-core>=0.3.61
langchain-community>=0.3.23
langchain-google-genai>=2.1.3
langgraph>=0.4.5
langchain-mistralai>=0.2.10
langchain-openai>=0.3.22

# External Service Modules
google-auth>=2.40.1
google-auth-oauthlib>=1.2.2
google-auth-httplib2>=0.2.0
google-api-python-client>=2.170.0
groundx>=2.3.0
openmeteo-requests>=1.5.0
requests-cache>=1.2.1
retry-requests>=2.0.0

# Utility modules
numpy>=2.2.6
pandas>=2.2.3
python-dotenv>=1.1.0
python-dateutil>=2.9.0
pydantic>=2.11.5
matplotlib>=3.10.3
